s/numpy.float/float/

+++ pyOpt/pyALGENCAN/pyALGENCAN.py
@@ -372,12 +372,12 @@ class ALGENCAN(Optimizer):
 		# Setup argument list values
 		nn = numpy.array([n], numpy.int)
 		mm = numpy.array([m], numpy.int)
-		lm = numpy.zeros([m], numpy.float)
+		lm = numpy.zeros([m], float)
 		coded = numpy.array([False,False,False,False,False,False,True,True,False,False], numpy.bool)
-		epsfeas = numpy.array([self.options['epsfeas'][1]], numpy.float)
-		epsopt = numpy.array([self.options['epsopt'][1]], numpy.float)
-		efacc = numpy.array([self.options['efacc'][1]], numpy.float)
-		eoacc = numpy.array([self.options['eoacc'][1]], numpy.float)
+		epsfeas = numpy.array([self.options['epsfeas'][1]], float)
+		epsopt = numpy.array([self.options['epsopt'][1]], float)
+		efacc = numpy.array([self.options['efacc'][1]], float)
+		eoacc = numpy.array([self.options['eoacc'][1]], float)
 		checkder = numpy.array([self.options['checkder'][1]], numpy.bool)
 		iprint = numpy.array([self.options['iprint'][1]], numpy.int)
 		if (myrank != 0):
@@ -390,9 +390,9 @@ class ALGENCAN(Optimizer):
 		if (iprint >= 0):
 			if os.path.isfile(ifile):
 				os.remove(ifile)
-		cnormu = numpy.array([0], numpy.float)
-		snorm = numpy.array([0], numpy.float)
-		nlpsupn = numpy.array([0], numpy.float)
+		cnormu = numpy.array([0], float)
+		snorm = numpy.array([0], float)
+		nlpsupn = numpy.array([0], float)
 		inform = numpy.array([0], numpy.int)
 
 		# Run ALGENCAN
